do not use Modular<int32_t> when getMaxModulus() assertion will fail

+++ trunk/linbox/linbox/algorithms/rational-solver-adaptive.h
@@ -39,7 +39,8 @@ namespace LinBox
 		static SolverReturnStatus solveNonsingular(OutVector& num, typename IRing::Element& den, const DenseMatrix<IRing>& M, const InVector& b)
 		{
 			linbox_check ((M. rowdim() == M. coldim()) && (b.size() == M.rowdim()) && (num. size() ==M.coldim()));
-			typedef Modular<int32_t> Field;
+			// typedef Modular<int32_t> Field;
+			typedef Modular<double> Field;
 			RationalSolver<IRing, Field, RandomPrimeIterator, WanTraits> numerical_solver;
 			//RationalSolver<IRing, Field, RandomPrimeIterator, NumericalTraits> numerical_solver;
 			SolverReturnStatus ret;
